+ labelled input and styling

+++ src/components/DataFrame/Catch/DataFrame.service.ts
@@ -3,8 +3,10 @@ import FireBaseApp from "../../../firebase"
 
 
 export const deleteCatch = async (catchId: string) => {
-  const db = getFirestore(FireBaseApp);
-  console.debug('deleteing', catchId);
-  const result = await deleteDoc(doc(db, "catches", catchId));
-  return result;
+  if (catchId.length > 0) {
+    const db = getFirestore(FireBaseApp);
+    console.debug('deleteing', catchId);
+    const result = await deleteDoc(doc(db, "catches", catchId));
+    return result;
+  }
 }
